# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Back2 Newsletters",
        "description": "Documentation Swagger pour la gestion des newsletters et archives.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur Local"
        }
    ],
    "paths": {
...
        "/projects": {
            "get": {
                "tags": [
                    "Projets"
                ],
                "summary": "Liste tous les projets",
                "operationId": "07f276c6e629646d087ee4b728e1234c",
                "responses": {
                    "200": {
                        "description": "Liste des projets",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Project"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Projets"
                ],
                "summary": "Créer un nouveau projet",
                "operationId": "d91af95183d5934b42aec1325d5165db",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "en cours",
                                        "enum": [
                                            "realise",
                                            "en_cours",
                                            "a_venir"
                                        ]
                                    },
                                    "image_path": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Projet créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/projects/{project}": {
            "get": {
                "tags": [
                    "Projets"
                ],
                "summary": "Détails d'un projet",
                "operationId": "a8d796d6ef9e7ecc37934df870dc68cf",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Projet trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Projet non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Projets"
                ],
                "summary": "Mettre à jour un projet",
                "operationId": "6d9ff23b79a456a9d18a3e2d6ebb73f1",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "en cours",
                                        "enum": [
                                            "realise",
                                            "en_cours",
                                            "a_venir"
                                        ]
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Projet mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Projets"
                ],
                "summary": "Supprimer un projet",
                "operationId": "31b7a0bd7545f2fe2f768180ffe4f15b",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Projet supprimé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Project": {
                "title": "Project",
                "description": "Modèle représentant un projet de l'organisation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Nom du projet"
                    },
                    "slug": {
                        "type": "string",
                        "example": "nom-du-projet"
                    },
                    "description": {
                        "type": "string",
                        "example": "Description détaillée du projet..."
                    },
                    "status": {
                        "type": "string",
                        "example": "en cours"
                    },
                    "image_path": {
                        "type": "string",
                        "example": "projects/image.png",
                        "nullable": true
                    },
                    "full_image_url": {
                        "type": "string",
                        "example": "http://localhost:8000/storage/projects/image.png",
                        "nullable": true
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "auteur": {
                        "description": "ID de l'utilisateur ayant créé le projet",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}